added some structure to patching
