+++ scoper.inc.php
@@ -2,7 +2,10 @@
 
 declare(strict_types=1);
 
+require_once 'scoper-fixes/twigPatcher.php';
+
 use Isolated\Symfony\Component\Finder\Finder;
+use function JdsDemoPlugin\Patchers\twigPatcher;
 
 return [
     // The prefix configuration. If a non null value will be used, a random prefix will be generated.
@@ -51,70 +54,7 @@ return [
     //
     // For more see: https://github.com/humbug/php-scoper#patchers
     'patchers' => [
-        function (string $filePath, string $prefix, string $contents): string {
-            try {
-                $twigFunctionsReferenceFile = realpath(__DIR__ . '/scoper-fixes/twigFunctions.json');
-
-                if ($twigFunctionsReferenceFile === false) {
-                    echo "Unable to determine the twig global function information file path: $twigFunctionsReferenceFile\n\n";
-                    die;
-                }
-
-                $twigRoot = join(DIRECTORY_SEPARATOR, ['twig', 'twig', 'src']) . DIRECTORY_SEPARATOR;
-
-                // only process twig files
-                $isTwigFile = false !== mb_strpos($filePath, $twigRoot);
-
-                if (!$isTwigFile) {
-                    return $contents;
-                }
-
-                $rawFunctionsFileContents = file_get_contents($twigFunctionsReferenceFile);
-
-                if ($rawFunctionsFileContents === false) {
-                    echo "Unable to read twig global function information: $twigFunctionsReferenceFile\n\n";
-                    die;
-                }
-
-                $segments = explode(DIRECTORY_SEPARATOR, $filePath);
-                $fileName = array_pop($segments);
-
-                $twigRootFunctionsLookup = json_decode($rawFunctionsFileContents, true);
-
-                // filter the complete list to only include functions not defined in this particular file
-                $twigRootFunctionsList = array_filter(
-                    array_keys($twigRootFunctionsLookup),
-                    fn (string $func) => $twigRootFunctionsLookup[$func] !== $fileName
-                );
-
-                $twigRootReplacements = array_map(
-                    fn ($func) => '\\' . $prefix . '\\' . $func,
-                    $twigRootFunctionsList
-                );
-
-                // replace global functions
-                $contents = str_replace($twigRootFunctionsList, $twigRootReplacements, $contents);
-
-
-                // todo replace string'd functions
-                $singleQuoteFunctionStrings = array_map(fn ($func) =>"'$func'", array_keys($twigRootFunctionsLookup));
-                $singeQuoteReplacements = array_map(fn ($func) => "'$prefix\\$func'", array_keys($twigRootFunctionsLookup));
-                $contents = str_replace($singleQuoteFunctionStrings, $singeQuoteReplacements, $contents);
-
-                // replace compilation strings
-                $moduleNode = 'ModuleNode.php';
-                if ($moduleNode === $fileName) {
-                    $contents = preg_replace_callback(
-                        "/\"use (Twig\\\+[^\"]+)\"/",
-                        fn (array $matches) => "\"use $prefix\\\\$matches[1]\"",
-                        $contents
-                    );
-                }
-            } catch (\Exception $e) {
-                echo $e;
-            }
-            return $contents;
-        },
+        JdsDemoPlugin\Patchers\twigPatcher()
     ],
 
     // PHP-Scoper's goal is to make sure that all code for a project lies in a distinct PHP namespace. However, you
